pyproject.toml
[project]
name = "videoaudiosync"
version = "0.1.0"
description = "Auto-sync clean audio to phone-recorded performance video, with editor"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "starlette>=0.40",
    "itsdangerous>=2.2",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "greenlet>=3.1",
    "alembic>=1.13",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "argon2-cffi>=23.1",
    "librosa>=0.10",
    "soundfile>=0.12",
    "numpy>=1.26,<2.2",
    "scipy>=1.14",
    "click>=8.1",
    "anyio>=4.6",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[project.scripts]
vasync = "app.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
